refactor api/types to fit graph teams structure/naming
