+++ src/components/BatchCardLarge/BatchCardLarge.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Card, CardBody, CardFooter, CardHeader, Col, Row } from 'reactstrap';
-import { Batch, batchSkillToImage } from '../../util/viewInfoSquadDummyTypes';
+import { Batch, batchSkillToImage } from '../../types';
 import InterventionModal from '../PlanInterventionModal/PlanInterventionModalv2';
 import mockDonutGraph from '../../assets/mock-donut-graph.png';
 import '../../scss/batch-card-large.scss';
